requirements.txt
httpx==0.28.1
playwright==1.51.0
beautifulsoup4==4.13.3
python-dotenv==1.1.0
